Test Structure Overhaul - Stage 1 - SM
