+++ tests/SMTests/smAbstractVerifier.py
@@ -1,0 +1,32 @@
+from algmatch.stableMarriageProblem import StableMarriageProblem
+
+from tests.abstractTests.abstractVerifier import AbstractVerifier
+from tests.SMTests.utils.instanceGenerator import SMInstanceGenerator
+from tests.SMTests.utils.minmaxSMs import MMSMS
+
+class SMAbstractVerifier(AbstractVerifier):
+    def __init__(self, total_men, total_women,
+                 lower_bound, upper_bound):
+        """
+        It takes argument as follows (set in init):
+            number of men
+            number of women
+            lower bound of the preference list length
+            upper bound of the preference list length
+        """
+
+        self._total_men = total_men
+        self._total_women = total_women
+        self._lower_bound = lower_bound
+        self._upper_bound = upper_bound
+
+        generator_args = (total_men, total_women,
+                          lower_bound, upper_bound)
+
+        AbstractVerifier.__init__(
+            self,
+            StableMarriageProblem,
+            SMInstanceGenerator,
+            generator_args,
+            MMSMS
+        )
