+++ tests/SMTests/utils/minmaxSMs.py
